More changes as pwer requests.
