+++ partner_contact_in_several_companies/models/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import multi_contact
+from . import res_partner, ir_actions
